agentic_backend/integration_layer/llm_client.py
import os
from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()


class LLMClient:
    def __init__(self):
        self.client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))

    def get_response(self, message: str) -> str:
        response = self.client.chat.completions.create(
            model="gpt-3.5-turbo",
            messages=[
                {
                    "role": "system",
                    "content": "You are a friendly AI companion for elderly users.",
                },
                {"role": "user", "content": message},
            ],
        )
        return response.choices[0].message.content.strip()
